backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
gspread==5.12.0
google-auth==2.23.4
python-dotenv==1.0.0
